requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
httpx==0.27.0
beautifulsoup4==4.12.3
scikit-learn==1.5.2
numpy==1.26.4
groq==0.5.0
python-dotenv==1.0.1
